Big changes
Rewrote a lot of stuff so it should be faster now. Took out a lot of redundant code, renamed the file console.lua, and also added a monkey patch for print and log so it's easier to call those things. The file only gained about 30 lines, which came from the monkeypatches and whitespace, so it's roughly the same size, but we don't use Classic anymore since it doesn't need to be OOP. Less pointless checks and math in general. Haven't thorougly tested, so there might be some hidden bugs but seems to be good for now.