Rect work
Had to change rect so that even if it's being skipped, it still checks for point in pos. We need to know if the mouse is still hovering and update regardless of whether the callback is run.

Basically, skip now only skips the callback and click/starthover/hold checks, doesn't skip the endhover check.

Also increased width of titlebar button size sort of poorly by just adding in a 1.4 to each value. Not sure if it's worth abstracting