Window moving
Added some mt options for point math. Point + point and all that. I ended up not needing them in this instant but I'm keeping them because I could see needing to do that sort of math later down the line.

Update self.mouse to include global position and delta. Slight reorginzation

Rect was update for the reorg

Added a display field for the display the window is in, as well as a move field containing state and offset. State is whether the window is currently being moved and the offset is the local offset, basically where the mouse was locally at the beginning of the move.

This is all used to basically move the window to the global position minus the offset along with the relevant display.

We get the global pos and display from global mouse which is updated in update

Also get the local position from mousemoved

We don't update any other rects while the window is moving.

global mouse is now Mouse to not conflict with the passed mouse vars for callbacks