All border drag callbacks work, but!
Unfortunately it's too laggy to update the border every update loop, possibly due to how setMode/updateMode work, possibly just because resizing involves recalculating a bunch of shit I have no idea. I thought of a couple solutions but they're all a bit overenginnered. As of right now, there is no visual to the window being expanded, it just pops into place on release of mouse. It would be nice to update that at some point.