Border resize work
Border resizing is now smooth! This is due to the window_manipulation file we added, which uses sdl2's windo manipulation functions to resize instead of love's. We also have two not technically callbacks, borderInit and borderMath. init calculates diff, and on the first loop saves the inital_win and stuff internally. borderMath is the checks for whether the window needs to move'd or resized. We also call console:resize directly here, to update our window while resizing. We however, call love.resize manually only when the resize process is finished, so that way it's more similar to how love actually does it. Otherwise, each border callback just calculates the new.x/y/w/h, places them into a table as well as gets the old (read; current) x/y/w/h. This is because if we are clicked on the border and not moving our mouse, there is no reason to be calling the resize/move function a bajillion times a second.

WE ARE NO LONGER CALCULATING WINDOW HEIGHT AND TITLEBAR SEPERATELY. Window height now accurately reflects the height of the window, which includes the titlebar height. This will make resizing a less painless process. I had made adescion in the beginning and stuck with it, and while internal window height is nice, I should save it as INTERNAL. I was clearly calculating external window height and saving internal and I noticed I was having to do weird workarounds for it. None of that.