Some more callback work
Been working on callback stuff. Have everything clickable done except for up/down arrow. I left those because they should control the line index, not the offset of the scrollbar, which is only  a visual representation of the index. However, I have done some work on the arrows as well as the scrollbar itself to calculate all of the offset stuff, to which we can calculate the index from backwards.

Also, the c:resize is now called when the window is being resized so that it doesn't look janky af