Input/Output text work
Added in a cursor, currently only chills at the end of the line. No arrow control yet

Added in basic typing, no backspace or anything

Added in output text, made it so that only what is visibly showing is what is visibly showing. No printing stuff offscreen to save on cycles