Slight adjustment on output text
I was only drawing below the certain index, but that meant scrolling up high enough was basically making everything beneath visible as well. Made it so it only ever drew a certain maximum amount, full_line_amount. It should never draw more than that amount of lines in one go, and that is calculated automatically at start and on window resize