Recalc line size
Made lineObj and recalc functions local, fixed scrollbar zooming off during resize,
Working on recalc line size for the internal size of a line measurement being wrong after resize

The idea is to have a small check, like a timestamp check or something, that we check per visible line, and if we need to, we recalculate the width/height of visible lines. So max it's only doing like 50 at a time, at full screen, regardless of how many lines there actually are.