Line work
So, due to newlines not really being handled correctly with the cursor, I took a moment to think about what I wanted. First of all, I had some random free floating functions which only were for lines, so I decided to make lines an OO thing. Secondly, there was no reason to have multiline input, because what is multiline input but multiple individual lines. But that meant that parsing text had to be done differently since there was always a chance of a line getting created that had newline characters in them, so I created a "multiLine" constructor that spits out Line objects, as many as needed. And since I'm not doing one string with line breaks anymore, the weird way I was calculating width was not neccassry. So I got rid of that. Then I realized I didn't even need width a as field, just as a var to calc height. That let me simplfiy parsing text and massively resimplify recalcSize