All sorts of things
Fixed multiline metatable. I didn't even realize that I was typing Line:new() everytime instead of just Line()

Added a text setter for SingleLine. Takes strings, tables, numbers, Lines, whatever, and changes itself to that text. Removes color formatting if you set one Line to another, use clone for that.

Added a clone

Fixed rounding on recalc size

Added a :len() function for Line.

Added a "L:" in front of any string'd line; I kept getting confused when it was outputting plain text not realizing it was an obj the whole time

Added a __type to Point, Rectangle, Line, and Color

Monkeypatch to type; if available on a table, it will check the __type metamethod.

Added some "do"'s so that I can fold some stuff away, like the callbacks and the basic functions and what not.

Working on enter/print right now