Took so many hours to get this working
I tried to use non breaking spaces but encoding is all janky. Honestly, that'll be a problem down the line but until then :shrug:. I've gotta handle it somehow but I'm not sure if I want to strip invalid char or import a library or rework how string works or what. This is the solution for now, and if it immediately becomes a problem I'll just strip any characters that'll break things but that's not my favorite solution.