Maximizing works correctly
Maximizing before would cover the task bar. Using SDL_GetDisplayUsableBounds, we can get the valid amount of space and only resize to that. A flag allows us to keep track of this.

Display can now sleep while app is running. Small qol

Rewrote the update to be slightly drier. Due to how order is important, it looks a bit funky but there's no repition